refactor: extract BuildingPlacementTools.cs + unified PlaceShape
